only get new distributions

+++ trunk/multidoc-jnr/getdoc.sh
@@ -8,11 +8,18 @@ do
   name=`echo $name | sed 's/\([^0-9]*\).*/\1/' | sed 's/-$//'`
   dest=$OUTPUT/$name/$version
 
-  mkdir -p $dest
-  unzip -q -d $dest $i *apidocs* 2>&1 | sed "s/caution: filename not matched:.*/No Javadoc in $name $version/"
-  mv $dest/*/docs/apidocs/* $dest 2>&1 | grep -v 'No such file or directory'
-  rmdir $dest/*/docs/apidocs 2>&1 | grep -v 'No such file or directory'
-  rmdir $dest/*/docs/ 2>&1 | grep -v 'No such file or directory'
-  rmdir $dest/*/ 2>&1 | grep -v 'Directory not empty' | grep -v 'No such file or directory'
-  rmdir $dest/ 2>&1 | grep -v 'Directory not empty' | grep -v 'No such file or directory'
+  if [ -d $dest ]
+  then
+    echo "Skipping $dest"
+  else
+    echo "Grabbing $dest"
+    mkdir -p $dest
+    unzip -q -d $dest $i *apidocs* 2>&1 | sed "s/caution: filename not matched:.*/No Javadoc in $name $version/"
+    mv $dest/*/docs/apidocs/* $dest 2>&1 | grep -v 'No such file or directory'
+    rmdir $dest/*/docs/apidocs 2>&1 | grep -v 'No such file or directory'
+    rmdir $dest/*/docs/ 2>&1 | grep -v 'No such file or directory'
+    rmdir $dest/*/ 2>&1 | grep -v 'Directory not empty' | grep -v 'No such file or directory'
+    rmdir $dest/ 2>&1 | grep -v 'Directory not empty' | grep -v 'No such file or directory'
+    rmdir $OUTPUT/$name | grep -v 'Directory not empty' | grep -v 'No such file or directory'
+  fi
 done
